+++ libythtbbs/misc.h
@@ -10,9 +10,9 @@ struct mymsgbuf {
 };
 void newtrace(char *s);
 int init_newtracelogmsq();
-int u2g(char *inbuf,int inlen,char *outbuf,int outlen);
+int u2g(char *inbuf,size_t inlen,char *outbuf,size_t outlen);
 int g2u(char *inbuf,size_t inlen,char *outbuf,size_t outlen);
-int code_convert(char *from_charset,char *to_charset,char *inbuf,int inlen,char *outbuf,int outlen);
+int code_convert(char *from_charset,char *to_charset,char *inbuf,size_t inlen,char *outbuf,size_t outlen);
 int is_utf_special_byte(unsigned char c);
 int is_utf(char * inbuf, size_t inlen);
 #endif
